feat: milestone 1.5 without getBalance (#286)
* wip: milestone 1.5 fixes

* feat: working 1.5 e2e transfer

* refactor: small cleanup

* fix: use actual paths in simulator oracle

* fix: nullifying aux data

* refactor: less magic numbers in byte count

* refactor: some refactors

+++ yarn-project/aztec-rpc/src/account_state/account_state.test.ts
@@ -91,7 +91,7 @@ describe('Account State', () => {
     expect(addTxAuxDataBatchSpy).toHaveBeenCalledWith([
       expect.objectContaining({
         ...ownedTxAuxDatas[0],
-        index: 2,
+        index: 2n,
       }),
     ]);
   });
@@ -116,15 +116,15 @@ describe('Account State', () => {
     expect(addTxAuxDataBatchSpy).toHaveBeenCalledWith([
       expect.objectContaining({
         ...ownedTxAuxDatas[0],
-        index: KERNEL_NEW_COMMITMENTS_LENGTH + 1,
+        index: BigInt(KERNEL_NEW_COMMITMENTS_LENGTH + 1),
       }),
       expect.objectContaining({
         ...ownedTxAuxDatas[1],
-        index: KERNEL_NEW_COMMITMENTS_LENGTH * 4,
+        index: BigInt(KERNEL_NEW_COMMITMENTS_LENGTH * 4),
       }),
       expect.objectContaining({
         ...ownedTxAuxDatas[2],
-        index: KERNEL_NEW_COMMITMENTS_LENGTH * 4 + 2,
+        index: BigInt(KERNEL_NEW_COMMITMENTS_LENGTH * 4 + 2),
       }),
     ]);
   });
